+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added a setting to enable the new solver without having to tinker with FFlags: `luau-lsp.fflags.enableNewSolver`.
+
 ### Changed
 
 - `workspace/diagnostic` call is ~98% faster when `luau-lsp.diagnostics.workspace` is disabled on large projects ([#826](https://github.com/JohnnyMorganz/luau-lsp/issues/826))